loader/requirements.txt
python-dotenv==0.19.2
requests==2.26.0
psycopg2==2.9.2
SQLAlchemy==1.4.27
configparser==5.2.0
PyMySQL==1.0.2
pandas==1.3.4
mongoengine==0.23.1
pymongo==3.11.3
numpy==1.21.4
Flask==2.0.2
minio==7.1.2
